Reverted changed
Reverted the previous changes, because it's redundant.
I don't need to check if a "player projectile" has hit a "enemy" or the "player" (friendly fire), because I've already configured the physics engine accordingly. A player projectile won't collide with a player in the first place. Same with the enemies. And a collision is a trigger for the attack. No clollision = no attack