#31 attacking turrets (#33)
* Turrets can attack

Different throw styles possible.
Currently:
* Horizontal Throw
* Parabolic Throw

* Adjusted masses

Soldiers and turret projectiles have different masses now, i.e. soldiers are getting knock-backed differently and according to the mass of the projectile.

* Turrets target enemies and deal damage

* Prepared logic to use the same code for enemy turrets too

* Reverted changed

Reverted the previous changes, because it's redundant.
I don't need to check if a "player projectile" has hit a "enemy" or the "player" (friendly fire), because I've already configured the physics engine accordingly. A player projectile won't collide with a player in the first place. Same with the enemies. And a collision is a trigger for the attack. No clollision = no attack

* Fixed NPE and minor other stuff

* Added "TargetShot" component

Adding this component to a projetile will cause a movement behavior like a missle (lock-on target while flying).